Let nodes know of their parents, poor orphans
